+++ schema/deleteLogSchema.js
